Create alternating Feeds / Rules view and clean settings tabs
