+++ src/components/RSS/Feeds/FeedIcon.vue
